hw 4.4 student service logs

+++ src/main/java/ru/hogwarts/school/service/StudentService.java
@@ -1,5 +1,7 @@
 package ru.hogwarts.school.service;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.hogwarts.school.model.Faculty;
@@ -8,11 +10,13 @@ import ru.hogwarts.school.repository.StudentRepository;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 public class StudentService {
     private final StudentRepository studentRepository;
+    private final static Logger logger = LoggerFactory.getLogger(StudentService.class);
 
     @Autowired
     public StudentService(StudentRepository studentRepository) {
@@ -20,33 +24,72 @@ public class StudentService {
     }
 
     public Student createStudent(Student student) {
-        return studentRepository.save(student);
-    }
-
-    public Student findStudent(Long id) {
-        return studentRepository.findStudentById(id);
-    }
-
-    public Student updateStudent(long id, Student student) {
-        Student updatedStudent = studentRepository.findStudentById(id);
-        if (updatedStudent == null) {
+        try {
+            logger.info("Создание нового студента: {}", student);
+            Student createStudent = studentRepository.save(student);
+            logger.info("Студент успешно создан. ID нового студента: {}", createStudent.getId());
+            return createStudent;
+        } catch (Exception e) {
+            logger.error("Ошибка при создании нового студента.", e);
             return null;
         }
-        updatedStudent.setName(student.getName());
-        updatedStudent.setAge(student.getAge());
-        return studentRepository.save(updatedStudent);
+    }
+
+    public Optional<Student> findStudent(Long id) {
+        logger.info("Trying to find student with ID: {}.", id);
+        Optional<Student> foundedStudent = studentRepository.findStudentById(id);
+        if (foundedStudent.isPresent()) {
+            logger.info("Student found with ID {}: {}", id, foundedStudent.get());
+        } else {
+            logger.info("Student not found with ID: {}", id);
+        }
+        return foundedStudent;
+    }
+
+    public Optional<Student> updateStudent(long id, Student student) {
+        logger.info("Trying to refactor student with ID: {}", id);
+        Optional<Student> foundedStudent = studentRepository.findStudentById(id);
+        if (foundedStudent.isPresent()) {
+            Student existingStudent = foundedStudent.get();
+            existingStudent.setName(student.getName());
+            existingStudent.setAge(student.getAge());
+            Student updatedStudent = studentRepository.save(existingStudent);
+            logger.info("Student ID {} successfully updated {}.", id, student);
+            return Optional.of(updatedStudent);
+        } else {
+            logger.info("Student ID {} not found.", id);
+            return Optional.empty();
+        }
     }
 
     public void deleteStudent(Long id) {
-        studentRepository.deleteById(id);
+        logger.info("Request for delete student with ID {}.", id);
+        if (studentRepository.existsById(id)) {
+            studentRepository.deleteById(id);
+            logger.info("Student with ID {} successfully deleted.", id);
+        } else {
+            logger.warn("Faculty not found with ID: {}. Deletion failed.", id);
+        }
     }
 
     public Collection<Student> findByAgeBetween(int min, int max) {
-        return studentRepository.findByAgeBetween(min, max);
+        logger.info("Request for collection students with age between MIN {} & MAX {}.", min, max);
+        Collection<Student> students = studentRepository.findByAgeBetween(min, max);
+        if (students.isEmpty()) {
+            logger.info("No students found with age between {} and {}.", min, max);
+        } else {
+            logger.info("{} students found with age between {} and {}.", students.size(), min, max);
+        }
+        return students;
     }
 
-    public Faculty getStudentsFaculty(long id) {
-        return studentRepository.findStudentById(id).getFaculty();
+
+    public Optional<Faculty> getStudentsFaculty(long id) {
+        logger.info("Request to the faculty for a student with ID {}.", id);
+        Optional<Faculty> facultyOptional = studentRepository.findStudentById(id)
+                .map(Student::getFaculty);
+        logger.info("Returned faculty {}", facultyOptional.orElse(null));
+        return facultyOptional;
     }
 
     public Collection<Student> filterAge(int age) {
